Revert "Merge PR 3281"

+++ src/core/frontend/framebuffer_layout.h
@@ -21,7 +21,7 @@ struct FramebufferLayout {
      * Returns the ration of pixel size of the top screen, compared to the native size of the 3DS
      * screen.
      */
-    u16 GetScalingRatio() const;
+    float GetScalingRatio() const;
 };
 
 /**
