+++ src/core/settings.h
@@ -108,7 +108,7 @@ struct Values {
     // Renderer
     bool use_hw_renderer;
     bool use_shader_jit;
-    u16 resolution_factor;
+    float resolution_factor;
     bool use_vsync;
     bool use_frame_limit;
     u16 frame_limit;
